pyproject.toml
[tool.poetry]
name = "telegram_kaspi"
version = "0.1.0"
description = ""
authors = ["damir <damir22astana@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
python-dotenv = "^0.19.2"
requests = "^2.27.1"
aiogram = "^2.19"
SQLAlchemy = "^1.4.31"
uvicorn = "^0.17.5"
psycopg2-binary = "^2.9.3"
httpx = "^0.22.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
